// src/design-system/components/popover/popover.types.ts
export enum PopoverPosition {
  Top = "top",
  TopLeft = "top-left",
  TopRight = "top-right",
  Bottom = "bottom",
  BottomLeft = "bottom-left",
  BottomRight = "bottom-right",
  Left = "left",
  LeftTop = "left-top",
  LeftBottom = "left-bottom",
  Right = "right",
  RightTop = "right-top",
  RightBottom = "right-bottom",
}

export enum AnimationsType {
  FadeUp = "fade-up",
  FadeDown = "fade-down",
  FadeLeft = "fade-left",
  FadeRight = "fade-right",
  Scale = "scale",
  None = "none",
}

export enum PopoverTrigger {
  Click = "click",
  Hover = "hover",
  Focus = "focus",
}